+++ pype/errors.py
@@ -1,0 +1,8 @@
+# -*- coding: utf-8 -*-
+"""Pype custom errors."""
+
+
+class PypeError(Exception):
+    """Custom error."""
+
+    pass