+++ pype/exceptions.py
@@ -1,8 +1,0 @@
-# -*- coding: utf-8 -*-
-"""Pype custom exceptions."""
-
-
-class PypeException(Exception):
-    """Custom exception."""
-
-    pass