+++ CHANGELOG.md
@@ -1,8 +1,13 @@
 # Changelog
 
+## 0.7.0
+
+- Rename PypeException to PypeError
+- Update of all libraries
+
 ## 0.6.7
 
-- Hotfix for bug in dependency managemen
+- Hotfix for bug in dependency management
 
 ## 0.6.6